refactor(nsfw+container-host): remove services/libs dep by copying slackLog
